refactor(sync): move per-tool modules under sandbox-docker/src/sync/agents/ (R3)
Completes the co-location: every per-tool sync op now lives under sync/. agents.ts
→ sync/agents/skills.ts (the ~/.agents skills volume — 100% sync). claude.ts /
codex.ts / opencode.ts → sync/agents/{claude,codex,opencode}.ts as whole-module
moves (the verified-clean coupling means the tool's launch/attach/login rides
along with its volume/seed/mount sync fns rather than risking an interleaved
split of 1400–1600-line files). Behaviour-preserving; public index API unchanged.
Repointed create/lifecycle/shell-session/docker-provider/docker-sync/index +
claude-credentials + the moved files' imports + tests/vi.mock.

Now `sandbox-docker/src/sync/` is the discoverable home of the docker sync layer;
`sandbox-cloud/src/sync/` the cloud one (R1).

Gate: build + typecheck + lint green; docker 310 tests green.
