+++ packages/sandbox-docker/src/shell-session.ts
@@ -1,5 +1,5 @@
 import { execa } from 'execa';
-import { buildTermSafeTmuxExec, buildTmuxSessionArgs, CONTAINER_USER } from './claude.js';
+import { buildTermSafeTmuxExec, buildTmuxSessionArgs, CONTAINER_USER } from './sync/agents/claude.js';
 
 /** Default tmux session name for `agentbox shell` (the box's first shell). */
 export const DEFAULT_SHELL_SESSION = 'shell';
