+++ packages/sandbox-docker/test/codex.test.ts
@@ -6,7 +6,7 @@ import {
   DEFAULT_CODEX_SESSION,
   resolveCodexVolume,
   SHARED_CODEX_VOLUME,
-} from '../src/codex.js';
+} from '../src/sync/agents/codex.js';
 
 describe('resolveCodexVolume', () => {
   it('returns the shared volume name when isolate is false', () => {
